changeSaveFunction
